(Hopefully) finalized project
Put all functions that coded the Tu matrix into a single .h and .cpp file. Commented all functions and made some minor adjustments
