+++ output.txt
@@ -1,1 +1,1 @@
-1.05445 0.393866 -0.0628656 0.146248 0.81749 0.693405 -0.152608 -1.33101 
+5 3.33333 5 1.66667 3.33333 -0.166667 0.166667 0.166667
