+++ parse.h
@@ -11,4 +11,7 @@ char get_type(string line);
 
 float get_value(string line);  
 
+//finds the source node and compares to previous max 
+int find_max_source_node(string line, int max_node);
+
 #endif
